something wrong in the fold fulkerson
It seems that there is something wrong in the modifications. After running the fold fulkerson I noticed, in the previous version the flow from d0 to sink is maximized while in here d0 to sink has a flow of 1. p4 has flow of 1 to d1, even though previously it doensn't. p4 should have flow 0 to d1 and flow 1 to d0.