p1 has flow to d0. I think this is caused by the greediness of the algorithm.
Because it saw edge from p1 to d0 first it just flows through it without considering all other edges. I suggest in first run it augment the flow of vertices where it is connected to 1 edges, then it second run it augments the flow of vertices with 2 edges, and so on. The maximum amount of edges a vertices can have is len(preferences)/5 *2 Because maximum amount of d and c vertices is len(preferences)/5 and it can be connected to d0,c0,d1,c1,...